refactor: rename & move EmailService, EmailVerificationService
