+++ chaekit-spring/src/main/java/qwerty/chaekit/service/member/admin/AdminService.java
@@ -21,14 +21,14 @@ import qwerty.chaekit.dto.page.PageResponse;
 import qwerty.chaekit.global.enums.ErrorCode;
 import qwerty.chaekit.global.exception.BadRequestException;
 import qwerty.chaekit.global.exception.NotFoundException;
-import qwerty.chaekit.service.member.notification.EmailService;
+import qwerty.chaekit.service.util.EmailNotificationService;
 import qwerty.chaekit.service.util.S3Service;
 
 @Service
 @RequiredArgsConstructor
 public class AdminService {
     private final PublisherProfileRepository publisherRepository;
-    private final EmailService emailService;
+    private final EmailNotificationService emailNotificationService;
     private final S3Service s3Service;
     private final UserProfileRepository userRepository;
 
@@ -85,7 +85,7 @@ public class AdminService {
 
         // 승인 처리
         publisher.approvePublisher();
-        emailService.sendPublisherApprovalEmail(publisher.getMember().getEmail());
+        emailNotificationService.sendPublisherApprovalEmail(publisher.getMember().getEmail());
     }
 
     @Transactional
@@ -100,7 +100,7 @@ public class AdminService {
 
         // 거절 처리
         publisher.rejectPublisher();
-        emailService.sendPublisherRejectionEmail(publisher.getMember().getEmail(), request.reason());
+        emailNotificationService.sendPublisherRejectionEmail(publisher.getMember().getEmail(), request.reason());
     }
 
     private static Pageable getPageableOrderedByCreatedAt(Pageable pageable) {
